feat(jf-playback-reporting): extract JellyfinAPI base to abstract class
