+++ server/routes/auth.ts
@@ -1,4 +1,4 @@
-import JellyfinAPI from '@server/api/jellyfin';
+import JellyfinMainAPI from '@server/api/jellyfinMain';
 import PlexTvAPI from '@server/api/plextv';
 import { ApiErrorCode } from '@server/constants/error';
 import { MediaServerType, ServerType } from '@server/constants/server';
@@ -287,7 +287,11 @@ authRoutes.post('/jellyfin', async (req, res, next) => {
     }
 
     // First we need to attempt to log the user in to jellyfin
-    const jellyfinserver = new JellyfinAPI(hostname ?? '', undefined, deviceId);
+    const jellyfinserver = new JellyfinMainAPI(
+      hostname ?? '',
+      undefined,
+      deviceId
+    );
 
     const ip = req.ip;
     let clientIp;
@@ -392,7 +396,7 @@ authRoutes.post('/jellyfin', async (req, res, next) => {
       }
 
       // Create an API key on Jellyfin from this admin user
-      const jellyfinClient = new JellyfinAPI(
+      const jellyfinClient = new JellyfinMainAPI(
         hostname,
         account.AccessToken,
         deviceId
